added changes from commit in github: added annotation for Transaction and changed reports generators
