--- src/main/java/org/shavneva/familybudget/reports/TypeReportGenerator.java
@@ -1,10 +1,0 @@
-package org.shavneva.familybudget.reports;
-
-import org.shavneva.familybudget.entity.Transaction;
-
-import java.util.List;
-import java.util.Map;
-
-public interface TypeReportGenerator {
-    byte[] generateReport(List<Transaction> transactionList, Map<String, Double> balances);
-}
