+++ src/main/java/org/shavneva/familybudget/reports/impl/WordGenerator.java
@@ -1,9 +1,13 @@
 package org.shavneva.familybudget.reports.impl.docsGenerator;
 
+import lombok.AllArgsConstructor;
 import org.apache.poi.xwpf.usermodel.*;
 import org.openxmlformats.schemas.wordprocessingml.x2006.main.*;
 import org.shavneva.familybudget.entity.Transaction;
-import org.shavneva.familybudget.reports.TypeReportGenerator;
+import org.shavneva.familybudget.reports.ReportGenerator;
+import org.shavneva.familybudget.service.BalanceService;
+import org.shavneva.familybudget.service.impl.TransactionService;
+import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
 
 import java.io.ByteArrayOutputStream;
@@ -14,9 +18,27 @@ import java.util.List;
 import java.util.Map;
 
 @Service
-public class WordGenerator implements TypeReportGenerator {
+@AllArgsConstructor
+public class WordGenerator implements ReportGenerator {
+
+    private final TransactionService transactionService;
+    private final BalanceService balanceService;
+
     @Override
-    public byte[] generateReport(List<Transaction> transactions, Map<String, Double> balances) {
+    public MediaType getSupportedMediaType() {
+        return MediaType.parseMediaType("application/vnd.openxmlformats-officedocument.wordprocessingml.document");
+    }
+
+    @Override
+    public String getFileExtension() {
+        return "docx";
+    }
+
+    @Override
+    public byte[] generateReport(String username, String date, String currency) {
+        List<Transaction> transactions = transactionService.getTransactionsByUser(username, date);
+        Map<String, Double> balances = balanceService.calculateBalances(transactions, currency);
+
         if (transactions.isEmpty()) {
             throw new IllegalArgumentException("Список транзакций пуст, невозможно создать отчет.");
         }
