+++ src/main/java/org/shavneva/familybudget/security/TransactionSecurity.java
@@ -1,0 +1,2 @@
+package org.shavneva.familybudget.security;public class TransactionSecurity {
+}
